+++ website/.prettierrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  useTabs: false,
+  singleQuote: false,
+  printWidth: 100,
+};