+++ crates/cargo_utils/CHANGELOG.md
@@ -6,6 +6,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.25](https://github.com/MarcoIeni/release-plz/compare/cargo_utils-v0.1.24...cargo_utils-v0.1.25) - 2023-08-02
+
+### Other
+- update dependencies
+
+## [0.1.24](https://github.com/MarcoIeni/release-plz/compare/cargo_utils-v0.1.23...cargo_utils-v0.1.24) - 2023-07-25
+
+### Other
+- update dependencies
+
+## [0.1.23](https://github.com/MarcoIeni/release-plz/compare/cargo_utils-v0.1.22...cargo_utils-v0.1.23) - 2023-06-26
+
+### Other
+- update dependencies
+
+## [0.1.22](https://github.com/MarcoIeni/release-plz/compare/cargo_utils-v0.1.21...cargo_utils-v0.1.22) - 2023-06-09
+
+### Other
+- update dependencies
+
+## [0.1.21](https://github.com/MarcoIeni/release-plz/compare/cargo_utils-v0.1.20...cargo_utils-v0.1.21) - 2023-05-24
+
+### Other
+- update dependencies
+
 ## [0.1.20](https://github.com/MarcoIeni/release-plz/compare/cargo_utils-v0.1.19...cargo_utils-v0.1.20) - 2023-05-21
 
 ### Other